Update documentation and contract.
This makes it clear that room is expected to be an Identifier subclass.
It also updates the text to reflect what it's meant to be (rather than
stating it's the user's full name).

Person is also required, as returning a 'frm' without a person field
makes the bot throw an exception complaining its required.